refactor: rename user transformer
